% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max.R
\name{min_max}
\alias{min_max}
\title{Identify Minimum and Maximum Values}
\usage{
min_max(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector, matrix, or data frame from which to find the
minimum and maximum values.}

\item{na.rm}{A logical value indicating whether to remove \code{NA} values before
computation. Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed to methods, if applicable.}
}
\value{
A list containing the minimum and maximum values.
}
\description{
This function identifies the minimum and maximum values in a numeric dataset
}
\examples{
# Example usage
my_data <- c(3, 5, 2, 8, 1, NA)
result <- min_max(my_data)
print(result)
}
